Guesser now can get to the right answer. Still have issue where guesser sometimes keeps guessing an incorrect answer.

+++ src/lib.hpp
@@ -9,7 +9,7 @@
 
 std::array<Correctness, 5> check(std::string answer, std::string guess);
 
-size_t play(std::string answer, Guesser* G);
+int play(std::string answer, Guesser* G);
 
 void make_hash_map_from_dictionary(std::unordered_map<std::string, int>& map, std::string filepath);
